Fixed downloading and scanning for volumes starting with a number (#41)
* added test case for comics beginning with numbers

* made small change to series regex to allow this to pass.
also updated testcase.

* Update comic series details in test cases

The details of the 'Infinity Gauntlet' and '100 Bullets' comic series in the tests for the backend file handling were updated. This includes changing the issue range for both series. Changes were made to keep the tests in sync with the updated data specifications and enhance the test coverage.

---------

Co-authored-by: Casvt <88994465+Casvt@users.noreply.github.com>