Removed MD fetch skipping based on last_updated value from CV
Kapowarr would check the value of the 'date_last_updated' key from CV to determine if it should fetch metadata for the volume. If it has changed since the last time, fetch, otherwise don't. However, it turns out that this value isn't updated when issues are added to the volume (or edited/removed/etc.). So newly released issues wouldn't be pulled, added and downloaded which is quite a big problem. So I removed this system for now.