Fixed extracting zips and then converting files
Since we've moved to always spawning processes, new DB connections needs to be opened in the processes we spawn. When extracting zip files, we would have uncommitted writes to the DB that would lead to an error when converting the extracted files in a process (the process needs these writes). Fixed by committing before starting the process pool.