Made docs clear to use mapped docker folder for RF
Users running with Docker would often try to enter the path of the root folder as it is on the host machine when trying to add a root folder. This commit makes the docs bring extra attention to the fact that the mapped folder should be added, not the host folder.