Log uncaught exceptions to console and log file
Uncaught exceptions would be handled the default way, which means printed to stderr. This commit changes this to it being logged using the logger, resulting in it still being printed to stderr, but also to the log file. This means that the downloaded log file now also contains uncaught exceptions.