Generate filename without issue title on EF misreport
When applying the EFD algorithm to the generated filename, we don't get back the same issue number(s) as that we originally made the filename for. This probably means that the title of the issue is messing up the algorithm. E.g. the title of issue 4 is "Book 1", then EFD might think the file is for issue 1 instead of 4. Try a name without the title and see if that fixes it. If so, use it. If not, then give up and just use the original name.